+++ CHANGELOG.md
@@ -5,13 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.13] - 2023-02-16
+### Added
+- Added the ability to add fields (`WithFields`) to the error enhancing the error message with structured information.
+
 ## [1.0.12] - 2022-10-21
 ### Added
 - `WithIgnoreString` now also check the underlying error message for the given string.
 
 ## [1.0.11] - 2022-10-21
 ### Added
-- Added the ability to tag errors with a custom tags. Tags is a list of tags which helps to categorize the error.
+- Added the ability to categorize errors with a custom tags (`WithTags`) enhancing the error message tags.
 
 ## [1.0.10] - 2022-10-21
 ### Changed